requirements.txt
line-bot-sdk
flask
openpyxl
tensorflow
Pillow